requirements.txt
scipy
pycocotools
pytorch_pretrained_bert
timm
fvcore
